refactor(frontend/payment): add payment module
