+++ packages/frontend/src/components/molecules/index.ts
@@ -1,7 +1,6 @@
 import Searcher from './Searcher';
-import ProductForBuy from './ProductForBuy';
 import CartCard from './CartCard';
 import BasicInfo from './BasicInfo';
 import ListProducts from './ListProducts';
 
-export { Searcher, ProductForBuy, CartCard, BasicInfo, ListProducts };
+export { Searcher, CartCard, BasicInfo, ListProducts };
