+++ packages/frontend/src/components/pages/index.ts
@@ -1,8 +1,7 @@
 import Main from './Main';
-import Payment from './Payment';
 import Cart from './Cart';
 import User from './User';
 import UserProducts from './UserProducts';
 import UserCart from './UserCart';
 
-export { Main, Payment, Cart, User, UserProducts, UserCart };
+export { Main, Cart, User, UserProducts, UserCart };
